Add support for custom env variables in IstioCNI
Add the ability to set custom environment variable for IstioCNI pod.
