+++ chart/crds/sailoperator.io_istiocnis.yaml
@@ -1074,6 +1074,12 @@ spec:
                           The directory path within the cluster node's filesystem where network namespaces are located.
                           Defaults to '/var/run/netns', in minikube/docker/others can be '/var/run/docker/netns'.
                         type: string
+                      env:
+                        additionalProperties:
+                          type: string
+                        description: 'A `key: value` mapping of environment variables
+                          to add to the pod'
+                        type: object
                       excludeNamespaces:
                         description: List of namespaces that should be ignored by
                           the CNI plugin.
@@ -1197,7 +1203,7 @@ spec:
                             type: string
                         type: object
                       resource_quotas:
-                        description: The resource quotas configration for the CNI
+                        description: The resource quotas configuration for the CNI
                           DaemonSet.
                         properties:
                           enabled:
